speed measurement function changed to float

+++ functions.h
@@ -6,7 +6,7 @@
 
 void help(); //print using instructions
 void readCenterCoordinates(list <Annotation>::iterator itList, list <Annotation> annotations); //read and print coordinates of centered pount of bounding box
-int speedMeasurement(list <Annotation>::iterator itList, list <Annotation> annotations, list <Annotation>::iterator prev, char axis = 'x'); //count speed of object beetwen current and previous frame
+float speedMeasurement(list <Annotation>::iterator itList, list <Annotation> annotations, list <Annotation>::iterator prev, char axis = 'x'); //count speed of object beetwen current and previous frame
 void drawRectangle(list <Annotation>::iterator itList, list <Annotation> annotations, string path, int blue[], int green[], int red[]); //draw a bounding box on each frame on video
 int randomNumber(); //find random number from 0 - 255
 int countNumberOfObjects(list <Annotation>::iterator itList, list <Annotation> annotations); //cout number of different objects
